Display and use the pitch and volume data already present in burger JSON files
This does require non-standard properties to work correctly, but that's still probably fine.  Additionally, only vitrine pages need to be regenerated, as the data is already present.

Note that this does not include all of the pitch changes that happen ingame, only those defined in sounds.json.  For instance, the break and place sounds use different pitch values, but that happens in code and is not tied to the sound event.